+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2058,7 +2058,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01M2WGX97ZQCH78AXJ84DT0D2H"
+    id: Optional[str] = "01M2WZ2G0R73YDQRJ52Y0ZMASY"
 
     description: Optional[str] = None